Clarifying SystmOne data flows
Where GP practices are on S1 and CHIS or SAIS providers are too, data
for children at these practices is likely to just appear on their
systems as a result of the API integration in Mavis. However, where
children are at EMIS or Medicus practices, data will need to be
manually extracted from Mavis and sent to CHIS or uploaded to
SAIS EPRs.